Adds a ResourceMerger to the AndroidBuilder library.
This will allow us to manually merge various resource folders.

This supports having multiple non-overlay folders while detecting
duplicates, as well as applying overlays.

The merger then creates a single resource folder that can be
fed to aapt.

This merger should be incremental (not yet) and be as efficient as possible.

Change-Id: I9e9b32954668160f97ff12fa70cd6078386b51c6