Use the ResourceMerger to handle overlay.
This creates a merged res folder that is then fed to aapt.

Right now the merged folder does not have its image
crunched, nor is it incremental.

This means the build will be slower (previously crunching was
incremental). Next step is to change aapt to be able to crunch
single images and then make the merging itself incremental.

Details:

The processImage task is replaced by mergeResourceTask.
It is also used for library project that had to merge their
resources before bundling, so the previous copy task is removed
and the merge is done directly in the bundle folder.

The API for AndroidBuilder is changed to reflect those changes.
processResource now only takes one res folder which is meant to
be the merged one. The getResourceInputs on VariantConfig
is changed to return a list of ResourceSets (except not actual
ResourceSet since this would load the resources, so instead
a list of list of files).

Also changed build.gradle for the gradle plugin to have
the integration test first deploy the plugin locally in order
to use the right one.

Change-Id: Ide96f733be835d94518b831baee7ee77b9a61dec