Improve on-device test result reporting.
Use ddmlib to run and gather the test results. The
library provides a test runner that can read the output
of the tests running on the device and generate Junit compatible
XML result reports.

This is then fed to gradle's report generator to make a standard report.

Also change the run test task to also do the install/uninstall
so that we can run tests on all connected devices in parallel.

The existing install/uninstall tasks can probably be removed.

Change-Id: I8d3baa848923a399de78dbdc7fc25d5548f56be2