Fix the aggregate reports.
When running manually we don't actually need the report.

When running from jenkins (or other CI) we really don't need
the report either as the CI server will take care of it.

We do need to aggregate the results though but we can't have that
depend on the result of the test themselves or it won't run in
case of failure.

Therefore I removed all the dependency (and the custom report).
The way to run for a CI server is:

./gradlew check aggregateResults --continue

Change-Id: I968bfd8696994c9fbc05288cd78c0b9a00e5a846