Improvements to SigningConfig/BuildType initialization.
Moved the init code out of the main classes in Builder and into
the Dsl children classes.

This fixes some issues around sharing the SigningConfig.debug
object instead of creating a new disconnected one, so that
customizing it does update the BuildType.debug object.

This also improve error handling so that Builder doesn't
throw a RuntimeException but the Gradle plugin can (actually
a BuildException) if the debug keystore path cannot be found.
It's better that the client of the Builder library decides
what to do instead of the library throwing.

Also added debug signing config customization support in the
library plugin.

Finally added a initWith() method to both SigningConfig and
BuildType to allow easy duplication from the DSL.
Will probably need to do the same for ProductFlavor.

Change-Id: I931be800f42346899655d66631029b7f4acc42a2