Ensures debug keystores are created.
For Each SigningConfig object, create a ValidateSigningTask
that creates the keystore if:
- the path match the default debug keystore
- the keystore is missing

Packaging tasks are made to depend on the validate task
matching the SigningConfig they are configured with.

The main reason to do that versus doing the creation
inside the packaging app is that the way gradle
handles input file, the file has to exist.
The optional annotation let the input be null,
but not be a value that doesn't match an existing
file.

Also adds the store type in the SigningConfig as it's necessary for
both creating and loaded a keystore.

Finally, this fixes the API of the packaging a bit:
- pass the SigningConfig directly everywhere.
- have the Builder throw the actual exceptions instead of
  RuntimeException (so that the caller can decide what to do)
- Make SigningConfigDsl extend the getters to be able to
  annotate them with @Input
- Clean up the DebugKeyHelper vs. KeystoreHelper.

Change-Id: If5eda8e2c2e759ace393d0689fa76699e3a9b6ba