Update the AIDL task to the new incremental framework.
Before, the aidl task would rely on the .d files created during
compilation to figure out what files needed recompilation.
The main drawback was that all .d files were created next to
the output files, making searching them before running the
compilation cumbersome.

This changes the mechanism to process the .d files right
after the compilation (when we know where to find them),
process them into a DependencyData object and later store
all those objects into a single file through DependencyDataStore.

Before the task run, it is much simpler to reload all the
DependencyData objects from the single file.

This now also relies on the IncrementalTask which provides
the exact list of changed files so this can be compared
to the DependencyData to know what files to recompile
and which output to update/remove.

Compilation of multiple aidl files now also use a
WaitableExecutor to run them in parallel.

Misc fix to dex task to ensure incremental command
receives all the parameters that are needed.

Change-Id: I3c2205d137072839c868decc4a98ae058f844948