Refactor tasks.
I previously split the tasks into an internal class
and a non-internal class to separate the public
and private api of the tasks, but all this did
was introduce complication for a split that
was not enforced at the DSL level anyway.

So this merges back the internal and non-internal
task classes.

Change-Id: Ib73cff2d6f3bc04e546a6f64844faaa8500ad2bb