Automate testing build output on devices.
This is to test the plugin itself, not a feature of the plugin.

'gradlew deviceTest' is similar to 'gradlew buildTest' in that
it requires an SDK through ANDROID_HOME, and then builds a set
of projects.

It differs in that it then pushed the project (and its test project)
to a connected device and runs the test there.

Note: This is even longer running that buildTest, and since it
requires a device, it is not a dependence on check. This must
be run separately.

Change-Id: I9d211ef22a680a4fba62314e943de3fdd54e4eee