Fix package renaming.
The previous mechanism used --rename-manifest-package while letting
the created manifest have the old package name. While this rename
the package entry in the manifest, the resource pool was still
associated to the original package name which made thinkgs awkward.

The new mechanism uses the new feature of the manifest-merger that allows
updating the package name. This means the application is compiled
completely using the new package.
In order to keep the R class sane though, aapt is called with
--custom-package to override only the location of the R class.

Change-Id: I0162dab6ca1cdeaf40e0c92d351f0290f01fd216