Only package local dependencies in builder.
the builder library previously packaged all its
dependencies in its jar. This started breaking
when bouncycastle was added due some issue in the
manifest.

This solves it but also does something we should have
done a while ago.

Now only the local dependencies (mostly tools/base) are
packaged in builder, while the external dependencies are
now declared as dependencies. Since the builder is pushed
to MavenCentral the dependencies can be found there.
(Sample projects find them in the local repo clone).

Change-Id: I789a2950056d0fe49c5f8fbaa11b149cc75a6361