Uses Set instead of List for VariantConfiguration.getPackagedJars.
The final result of getPackagedJars depends on two collections: mJars
and mFlatLibraries. This means that if there are dependencies that
include the same jars, the method will return duplicate jar paths.

This is problematic for packaging tools like ProGuard because these
duplicated jar paths are added to injars, which will produce an error.

This patch fixes that by using a HashSet for collection and returns an
ArrayList of unique File objects.

Change-Id: I2bd250caf4a886107e2090941a2be416f3fc70a7
Signed-off-by: Jonathan Le <jle@twitter.com>