Improve model.
Since we are already changing the model in non compatible
way, now is a good time to fix some issues.

Refactoring all the properties of an artifact out of variant
and into its own ArtifactInfo class, so that we can have both
the main and test artifact in the variant in an elegant way.

Also adding SigningConfig info to enable the creation of
export wizard from the IDE.

Change-Id: Ib6c48873ba0e5c5a61bc1e79b00a90e2be83652f