Add support for library coming from modules.
This changes adds a "project identifier" to the AndroidLibrary
interface. this is only used in Gradle right now but is generic
enough for other usage of the model.

This also removes the previous AndroidLibraryProject interface as
the way the model is passed back in Gradle prevents using children
interfaces.

This is currently a hack based on the location of the bundle file
related to the output of the modules. This is temporary as we need
an API in Gradle to do it properly.

Change-Id: I51f0d501941a08125376d15cc83742a17730fadd